moving helper js function to helpers.js file
